Bidder - Get Started Update
Corrections to Getting Started as a Bidder. Also hid it from showing up as I believe it shouldn't be named Get Started since it is already under Get Started + it might be a better fit under Concepts instead of Get Started. Hid it until we figure out what to do with it.